requirements.txt
streamlit
google-generativeai
plotly
pandas
matplotlib
python-dotenv